Add base class for classification
